#50-Modified the MainMax class and the test class
