+++ java-algorithms-challenge/src/main/java/fr/afrogeek/algorithms/MaxOfTwo.java
@@ -1,7 +1,37 @@
 package fr.afrogeek.algorithms;
 
 public class MaxOfTwo {
-    public static int maxOfTwoNumbers(int a, int b){
-        return a>b?a:b;
-    }
+
+	private static int result;
+
+
+	public static void main(String[] args) {
+		
+    setResult(maxOfTwoNumbers(2,3));
+		
+	}
+	
+	
+	public static int maxOfTwoNumbers(int a, int b) {
+		
+		if (a > b) {
+			return a;
+		}
+		
+		else {
+			return b;
+		}
+		
+	}
+
+
+	public static int getResult() {
+		return result;
+	}
+
+
+	public static void setResult(int result) {
+		MaxOfTwo.result = result;
+	}
+
 }
